+++ src/app/admin/manage-colleges/manage-colleges.component.ts
@@ -1,14 +1,16 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnDestroy, OnInit } from '@angular/core';
 import { FormGroup, FormControl, Validators } from '@angular/forms';
+import { Subscription } from 'rxjs';
 import { College } from 'src/app/models/college';
 import { CollegeService } from 'src/app/services/college.service';
+import { ModalService } from 'src/app/services/modal.service';
 
 @Component({
   selector: 'app-manage-colleges',
   templateUrl: './manage-colleges.component.html',
   styleUrls: ['./manage-colleges.component.css']
 })
-export class ManageCollegesComponent implements OnInit {
+export class ManageCollegesComponent implements OnInit, OnDestroy {
 
   form: FormGroup;
   error: String = null;
@@ -16,8 +18,12 @@ export class ManageCollegesComponent implements OnInit {
   isLoading: boolean = false;
   private selectedCollege: College = null;
   colleges: College[] = [];
+  private modalSub: Subscription;
 
-  constructor(private collegeService: CollegeService) { }
+  constructor(
+    private collegeService: CollegeService,
+    private modalService: ModalService
+  ) { }
 
   ngOnInit(): void {
     this.form = new FormGroup({
@@ -47,25 +53,32 @@ export class ManageCollegesComponent implements OnInit {
   }
 
   onUpdate(){
-    const data : any = {
-      name: this.form.value.name,
-      regNo: this.form.value.regNo
-    }
-    this.isLoading = true;
-    this.collegeService.updateCollegeData(this.selectedCollege.cid, data)
+    this.modalSub = this.modalService.OpenConfirmCancelModal(ModalService.UPDATE_MESSAGE, false)
     .subscribe(
-      () => {
-        this.isLoading = false;
-        this.showSuccess = true;
-        setInterval(() => {this.showSuccess = false}, 2000);
-        this.fetchCollegeData();
-        this.selectedCollege = null;
-        this.form.reset();
-      },
-      err => {
-        this.isLoading = false;
-        this.error = err;
-        setInterval(() => {this.error = null}, 5000);
+      result => {
+        if(result){
+          const data : any = {
+            name: this.form.value.name,
+            regNo: this.form.value.regNo
+          }
+          this.isLoading = true;
+          this.collegeService.updateCollegeData(this.selectedCollege.cid, data)
+          .subscribe(
+            () => {
+              this.isLoading = false;
+              this.showSuccess = true;
+              setInterval(() => {this.showSuccess = false}, 2000);
+              this.fetchCollegeData();
+              this.selectedCollege = null;
+              this.form.reset();
+            },
+            err => {
+              this.isLoading = false;
+              this.error = err;
+              setInterval(() => {this.error = null}, 5000);
+            }
+          )
+        }
       }
     )
   }
@@ -80,21 +93,28 @@ export class ManageCollegesComponent implements OnInit {
   }
 
   onDelete(){
-    this.isLoading = true;
-    this.collegeService.deleteCollege(this.selectedCollege.cid)
+    this.modalSub = this.modalService.OpenConfirmCancelModal(ModalService.DELETE_MESSAGE, true)
     .subscribe(
-      () => {
-        this.isLoading = false;
-        this.showSuccess = true;
-        setInterval(() => {this.showSuccess = false}, 2000);
-        this.fetchCollegeData();
-        this.selectedCollege = null;
-        this.form.reset();
-      },
-      err => {
-        this.isLoading = false;
-        this.error = err;
-        setInterval(() => {this.error = null}, 5000);
+      result => {
+        if(result){
+          this.isLoading = true;
+          this.collegeService.deleteCollege(this.selectedCollege.cid)
+          .subscribe(
+            () => {
+              this.isLoading = false;
+              this.showSuccess = true;
+              setInterval(() => {this.showSuccess = false}, 2000);
+              this.fetchCollegeData();
+              this.selectedCollege = null;
+              this.form.reset();
+            },
+            err => {
+              this.isLoading = false;
+              this.error = err;
+              setInterval(() => {this.error = null}, 5000);
+            }
+          )
+        }
       }
     )
   }
@@ -102,6 +122,10 @@ export class ManageCollegesComponent implements OnInit {
   onClear(){
     this.form.reset();
     this.selectedCollege = null;
+  }
+
+  ngOnDestroy(){
+    if(this.modalSub) this.modalSub.unsubscribe();
   }
 
 }
