+++ src/app/admin/manage-spoc/manage-spoc.component.ts
@@ -1,6 +1,8 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnDestroy, OnInit } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
+import { Subscription } from 'rxjs';
 import { SPOC } from 'src/app/models/spoc';
+import { ModalService } from 'src/app/services/modal.service';
 import { SpocService } from 'src/app/services/spoc.service';
 
 @Component({
@@ -8,7 +10,7 @@ import { SpocService } from 'src/app/services/spoc.service';
   templateUrl: './manage-spoc.component.html',
   styleUrls: ['./manage-spoc.component.css']
 })
-export class ManageSpocComponent implements OnInit {
+export class ManageSpocComponent implements OnInit, OnDestroy {
 
   form: FormGroup;
   error: String = null;
@@ -16,8 +18,12 @@ export class ManageSpocComponent implements OnInit {
   showSuccess: boolean = false;
   private selectedSpoc: SPOC = null;
   spocs: SPOC[] = [];
+  private modalSub: Subscription;
 
-  constructor(private spocService: SpocService) { }
+  constructor(
+    private spocService: SpocService,
+    private modalService: ModalService
+  ) { }
 
   ngOnInit(): void {
     this.form = new FormGroup({
@@ -48,26 +54,33 @@ export class ManageSpocComponent implements OnInit {
   }
 
   onUpdate(){
-    const data : any = {
-      email: this.form.value.email,
-      name: this.form.value.name,
-      contact: this.form.value.contact
-    }
-    this.isLoading = true;
-    this.spocService.updateSpocData(this.selectedSpoc.spocId, data)
+    this.modalSub = this.modalService.OpenConfirmCancelModal(ModalService.UPDATE_MESSAGE,false)
     .subscribe(
-      () => {
-        this.isLoading = false;
-        this.showSuccess = true;
-        setInterval(() => {this.showSuccess = false}, 2000);
-        this.fetchSpocData();
-        this.selectedSpoc = null;
-        this.form.reset();
-      },
-      err => {
-        this.isLoading = false;
-        this.error = err;
-        setInterval(() => {this.error = null}, 5000);
+      result => {
+        if(result){
+          const data : any = {
+            email: this.form.value.email,
+            name: this.form.value.name,
+            contact: this.form.value.contact
+          }
+          this.isLoading = true;
+          this.spocService.updateSpocData(this.selectedSpoc.spocId, data)
+          .subscribe(
+            () => {
+              this.isLoading = false;
+              this.showSuccess = true;
+              setInterval(() => {this.showSuccess = false}, 2000);
+              this.fetchSpocData();
+              this.selectedSpoc = null;
+              this.form.reset();
+            },
+            err => {
+              this.isLoading = false;
+              this.error = err;
+              setInterval(() => {this.error = null}, 5000);
+            }
+          )
+        }
       }
     )
   }
@@ -83,27 +96,37 @@ export class ManageSpocComponent implements OnInit {
   }
 
   onDelete(){
-    this.isLoading = true;
-    this.spocService.deleteSpoc(this.selectedSpoc.spocId)
-    .subscribe(
-      () => {
-        this.isLoading = false;
-        this.showSuccess = true;
-        setInterval(() => {this.showSuccess = false}, 2000);
-        this.fetchSpocData();
-        this.selectedSpoc = null;
-        this.form.reset();
-      },
-      err => {
-        this.isLoading = false;
-        this.error = err;
-        setInterval(() => {this.error = null}, 5000);
+    this.modalSub = this.modalService.OpenConfirmCancelModal(ModalService.DELETE_MESSAGE, true)
+    .subscribe(result => {
+      if(result){
+        this.isLoading = true;
+        this.spocService.deleteSpoc(this.selectedSpoc.spocId)
+        .subscribe(
+          () => {
+            this.isLoading = false;
+            this.showSuccess = true;
+            setInterval(() => {this.showSuccess = false}, 2000);
+            this.fetchSpocData();
+            this.selectedSpoc = null;
+            this.form.reset();
+          },
+          err => {
+            this.isLoading = false;
+            this.error = err;
+            setInterval(() => {this.error = null}, 5000);
+          }
+        )
       }
-    )
+    })
   }
 
   onClear(){
     this.form.reset();
     this.selectedSpoc = null;
   }
+
+  ngOnDestroy(){
+    if(this.modalSub) this.modalSub.unsubscribe();
+  }
+  
 }
